+++ R/pr.R
@@ -475,6 +475,7 @@ pr_filter <- function(pr,
 #' @param port A number or integer that indicates the server port that should
 #' be listened on. Note that on most Unix-like systems including Linux and
 #' Mac OS X, port numbers smaller than 1025 require root privileges.
+#' @param quiet If `TRUE`, don't print routine startup messages.
 #'
 #' @examples
 #' \dontrun{
@@ -488,11 +489,13 @@ pr_filter <- function(pr,
 #' @export
 pr_run <- function(pr,
                    host = '127.0.0.1',
-                   port = getOption('plumber.port', NULL)
+                   port = getOption('plumber.port', NULL),
+                   quiet = FALSE
 ) {
   validate_pr(pr)
   pr$run(host = host,
-         port = port)
+         port = port,
+         quiet = quiet)
 }